requirements.txt
discord.py
openai
chromadb
sentence-transformers
beautifulsoup4
lxml
requests
python-dotenv